Refactor existing customizations, removing the ones no longer needed and using the new pattern in the others
